feature/router-improvement: improve router and add Login controller
