+++ App/Admin/Controller/NewTechnologyController.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace Admin\Controller;
+
+
+use Admin\Core\Config\AbstractController;
+
+class NewTechnologyController extends AbstractController
+{
+    public function index(){
+        var_dump('hello new techno');die;
+    }
+}
