+++ App/Admin/Resources/Views/admin_home.phtml
@@ -19,9 +19,9 @@
                     <p>Bienvenue dans le Back Office !!</p>
                 </div>
 
-                <form action="/admin/logout" method="get">
-                    <button type='submit'>DECONNEXION</button>
-                </form>
+                <div class="adminLogout">
+                    <a href="/admin/login/logout" title="logout">DECONNEXION</a>
+                </div>
             </div>
         </header>
         <main class="admin-wrapper flex">
@@ -33,7 +33,7 @@
                             <a href="/admin/newProject">Ajouter un projet</a>
                         </li>
                         <li>
-                            <a href="/admin/newTechnoogy">Ajouter une techno</a>
+                            <a href="/admin/newTechnology">Ajouter une techno</a>
                         </li>
                         <li>
                             <a href="/admin/newLangage">Ajouter un Langage</a>
