+++ App/Admin/Resources/Views/admin_login.phtml
@@ -20,7 +20,7 @@
             <h1>Admin login</h1>
         </header>
         <main>
-            <form action="/admin/login" method="post">
+            <form action="/admin/login/checkLogin" method="post">
                 <div>
                     <label for="login">Login</label>
                     <input id="login" name="login" type="text"/>
